Change cons_puthex to 64-bit. Add chekcing ELF magic.

+++ driver/tty.h
@@ -20,6 +20,6 @@ struct console {
 void cons_clear_screen(void);
 void cons_putchar(char ch);
 void cons_puts(const char *s);
-void cons_puthex(uint32_t hex);
+void cons_puthex(uint64_t hex);
 
 #endif
